Implement big number data structure
A big number data structure is implemented using linux-style list api.
Each node of the list stores up to 18 digits of number in decimal.
The head of the list stores the length of the list and the sign of
the big number.
The name of the module is changed in the Makefile
for using multiple source file to build the module.