Fix set function with large numbers
The original bn_set is called assuming the number of nodes is allocated
when calling bn_new, which may leads to error when the calculation is
incorrect. In this commit, bn_set will allocate a new node when the
number of nodes aren't enough. Since the argument lies in uint64_t, the
number of new node allocated in bn_set will be at most one.