Introduce swap to Fibonacci naive method
The original fibonacci naive uses bn_add_to_smaller to add the larger
number to the smaller one, then return the number corresponding to k,
which could suffer from branch miss since there are two "if" statements.
In this commit, the xor swap was introduce to the naive method in order
to decrease the number of "if"s in its implementation. However, two
"if"s are added to determine if the calculation is nedded.